backend/requirements.txt
# Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
praw==7.7.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4
xgboost==2.0.3
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
prophet==1.1.5
statsmodels==0.14.0
vaderSentiment==3.3.2
transformers==4.35.2
torch==2.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

